Complete assignment 5.

+++ assignment5/kmeans.py
@@ -1,5 +1,6 @@
 import sys
 
+import numpy as np
 import pandas as pd
 
 from libs import k_means
@@ -17,7 +18,7 @@ def main():
 
     arr = pd.read_csv(sys.argv[1], header=None).to_numpy()
     data = arr[:, 2:]
-    labels = arr[:, 1]
+    labels = arr[:, 1].astype(np.int_)
     centroids = k_means(data, int(sys.argv[2]))
     print(
         f'WC-SSD: {sum_of_within_cluster_squared_distances(centroids, data)}')
